Minor changes, mostly affecting windows builds

+++ src/RUVIIIC.cpp
@@ -201,7 +201,7 @@ Rcpp::RObject RUVIIIC(Rcpp::NumericMatrix input, int k, Rcpp::NumericMatrix M, R
 				//Create the view which has the correct number of rows
 				auto selectRowsFromInputView = selectRowsFromInput.block(0, 0, nSubmatrixRows, nRows);
 				selectRowsFromInputView = Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor>::Constant(nSubmatrixRows, nRows, 0);
-				for(int i = 0; i < nonMissingIndices.size(); i++)
+				for(int i = 0; i < static_cast<int>(nonMissingIndices.size()); i++)
 				{
 					selectRowsFromInputView(i, nonMissingIndices[i]) = 1;
 				}
